# apidocs.yml
openapi: '3.0.3'
info:
  title: API Title
  version: '1.0'
servers:
  - url: http://localhost:8080
    description: Local server
components:
  securitySchemes:
    regularSecurity:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
  /all:
    get:
      summary: Get all users
      security:
        - regularSecurity: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    name:
                      type: string
                    age:
                      type: integer
    post:
      summary: Get all users with a filter
      security:
        - regularSecurity: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                search:
                  type: string
                sort_by:
                  type: string
                order:
                  type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    name:
                      type: string
                    age:
                      type: integer
  /detail/{id}:
    get:
      summary: Get user detail
      security:
        - regularSecurity: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
                  age:
                    type: integer
  /update/{id}:
    put:
      summary: Update user
      security:
        - regularSecurity: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                age:
                  type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
                  age:
                    type: integer